refactor: split basket page into components
