+++ src/components/ProductCard/constants.ts
@@ -1,5 +1,3 @@
-export const centsInEuro = 100;
-
 export enum PageData {
   addTooltip = 'Add to Cart',
   addSnackbar = 'Product was added to cart',
