+++ src/core/commonConstants.ts
@@ -1,1 +1,5 @@
 export const lsUserKey = 'plantastic-user';
+
+export const centsInEuro = 100;
+
+export const currency = '€';
