+++ src/enums/basketRespResults.enum.ts
@@ -1,0 +1,5 @@
+export enum BasketRespResultMessages {
+  ITEM_REMOVED = 'Removed successfully',
+  PROMO_REMOVED = 'Promo code removed',
+  PROMO_APPLIED = 'Promo code applied',
+}
