+++ src/pages/Basket/constants.ts
@@ -1,3 +1,1 @@
 export const drawerWidth = '350px';
-export const centsPerEuro = 100;
-export const currency = '€';
